Update cheat sheet. Improve the commit message auto command.
Let readers know that Tab will both open and close the miniBufExplorer.

The auto command would display an error message if the branch name
didn't contain 8 or more digits to pull out. I added silent! to fix
that, and then added additional commands to remove the [#] in line 1
that would result from the Put keystroke.